Namespace all types by their API version
